feat: 优化代码的风格
