+++ src/divider/index.tsx
@@ -1,0 +1,1 @@
+export { Divider } from 'antd'
