+++ src/grid/index.tsx
@@ -1,0 +1,1 @@
+export { Row, Col } from 'antd'
