+++ src/index.ts
@@ -26,13 +26,13 @@ export { default as Form } from './form'
 export type { JSONFormsData } from './form'
 export { Image } from './image'
 export { default as InputNumber } from './input-number'
-export { default as Input } from './input/Input'
-export { Card } from './layout/Card'
-export { Divider } from './layout/Divider'
-export { Col, Row } from './layout/Grid'
-export { Layout } from './layout/Layout'
-export { Space } from './layout/Space'
-export type { SpaceProps } from './layout/Space'
+export { default as Input } from './input'
+export { Card } from './card'
+export { Divider } from './divider'
+export { Col, Row } from './grid'
+export { Layout } from './layout'
+export { Space } from './space'
+export type { SpaceProps } from './space'
 export { List } from './list'
 export { Menu } from './menu/Menu'
 export { message } from './message'
