--- src/layout/Card.tsx
@@ -1,1 +1,0 @@
-export { Card } from 'antd'
