--- src/layout/Divider.tsx
@@ -1,1 +1,0 @@
-export { Divider } from 'antd'
