--- src/layout/Layout.tsx
@@ -1,1 +1,0 @@
-export * as Layout from 'react-spaces';
