--- src/layout/Space.tsx
@@ -1,2 +1,0 @@
-export { Space } from 'antd'
-export type { SpaceProps } from 'antd'
