+++ src/layout/index.tsx
@@ -1,0 +1,1 @@
+export * as Layout from 'react-spaces'
