+++ src/space/index.md
@@ -15,14 +15,13 @@ group:
 
 例子:
 
-
 ```tsx
 /**
  * title: 基本
  * desc: 最简单的用法，在浮层中可以选择或者输入日期。
  */
-import React, { useState, useRef } from 'react';
-import { Space, Button } from '@weblif/fast-ui';
+import React, { useState, useRef } from 'react'
+import { Space, Button } from '@weblif/fast-ui'
 
 export default () => {
     return (
@@ -32,20 +31,18 @@ export default () => {
             <Button>按钮二</Button>
             <Button>按钮三</Button>
         </Space>
-    );
-};
+    )
+}
 ```
-
-
 
 ```tsx
 /**
  * title: 垂直间距
  * desc: 相邻组件垂直间距。
  */
-import React, { useState, useRef } from 'react';
-import { Space } from '@weblif/fast-ui';
-import { Card } from 'antd';
+import React, { useState, useRef } from 'react'
+import { Space } from '@weblif/fast-ui'
+import { Card } from 'antd'
 
 export default () => {
     return (
@@ -59,20 +56,18 @@ export default () => {
                 <p>Card content</p>
             </Card>
         </Space>
-    );
-};
+    )
+}
 ```
-
 
 <br />
 
 ### API
 
-
-| 属性           | 说明                                    | 类型              | 默认值 |
-| -------------- | --------------------------------------- | ----------------- | ------ |
-|align           |对齐方式                                  |`start` \| `end` \|`center` \|`baseline`
-|direction       |间距方向                                  |`vertical` \| `horizontal`
-|size            |间距大小                                  |`Size` | `Size[]` | 
-|split           |设置拆分                                  |`ReactNode`
-|wrap            |是否自动换行，仅在 `horizontal` 时有效       |`boolean`
+| 属性      | 说明                                   | 类型                                     | 默认值   |
+| --------- | -------------------------------------- | ---------------------------------------- | -------- |
+| align     | 对齐方式                               | `start` \| `end` \|`center` \|`baseline` |
+| direction | 间距方向                               | `vertical` \| `horizontal`               |
+| size      | 间距大小                               | `Size`                                   | `Size[]` |
+| split     | 设置拆分                               | `ReactNode`                              |
+| wrap      | 是否自动换行，仅在 `horizontal` 时有效 | `boolean`                                |
