+++ src/space/index.tsx
@@ -1,0 +1,2 @@
+export { Space } from 'antd'
+export type { SpaceProps } from 'antd'
